third_party/WebKit/LayoutTests/fast/canvas-api/canvas-getImageData-rounding.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
<script>
test(function(t) {
    ctx = document.createElement('canvas').getContext('2d');

    function testDimensions(sx, sy, sw, sh, width, height)
    {
        imageData = ctx.getImageData(sx, sy, sw, sh);
        assert_equals(imageData.width, width);
        assert_equals(imageData.height, height);
    }

    // Basic integer values
    testDimensions(   0,  0,   20,  10,     20, 10);

    // Source point is not an integer
    testDimensions(  .1, .2,   20,  10,     20, 10);
    testDimensions(  .9, .8,   20,  10,     20, 10);

    // Size is not an integer
    testDimensions(   0,  0, 20.9, 10.9,     20, 10);
    testDimensions(   0,  0, 20.1, 10.1,     20, 10);

    // Basic integer negative values
    testDimensions(  -1, -1,   20,  10,     20, 10);

    // Non-integer negative values
    testDimensions(-1.1,  0,   20,  10,     20, 10);
    testDimensions(-1.9,  0,   20,  10,     20, 10);
}, 'Test the handling of non-integer source coordinates in getImageData().');
</script>
</body>
